chore: moved LogLevel enum to data_structure
